use static folder for call and fix button css
