+++ client/app/main/main.html
@@ -97,7 +97,7 @@
             <dt>Friday (4:00-5:15)</dt>
             <dd>Large group meetings in DCC 318</dd>
           </dl>
-          <a class="btn btn-primary btn-right" ng-href="https://docs.google.com/spreadsheets/d/1A6naU2lG5ALS7RELFpeG__QHaUjLbwXmCGIyzGJum7s/edit#gid=0" role="button">Presentation Schedule</a>
+          <a class="btn btn-primary pull-right" ng-href="https://docs.google.com/spreadsheets/d/1A6naU2lG5ALS7RELFpeG__QHaUjLbwXmCGIyzGJum7s/edit#gid=0" role="button">Presentation Schedule</a>
         </div>
       </div>
       <!-- End Meeting Times Panel -->
