+++ client/app/main/main.less
@@ -44,7 +44,3 @@ body {
   margin-bottom: 90px;
   font-size: 16px;
 }
-
-.btn-right {
-  float: right;
-}
