+++ server/api/static/static.controller.js
@@ -1,0 +1,14 @@
+/**
+ * Static data controller
+ */
+
+'use strict';
+
+var config = require('../../config/environment');
+
+// Get email
+exports.email = function(req, res) {
+    var email = {};
+    email.address = config.contactEmail;
+    return res.status(200).send(email);
+}
