+++ server/api/static/static.spec.js
@@ -1,0 +1,2 @@
+'use strict';
+
