+++ server/routes.js
@@ -20,13 +20,14 @@ export default function(app) {
   app.use('/api/classyear', require('./api/classyear'));
   app.use('/api/smallgroup', require('./api/smallgroup'));
   app.use('/api/attendance', require('./api/attendance'));
+  app.use('/api/static', require('./api/static'));
 
   app.use('/uploads', express.static(config.imageUploadPath));
 
   app.use('/auth', require('./auth'));
 
   // All undefined asset or api routes should return a 404
-  app.route('/:url(api|auth|components|app|bower_components|assets|uploads)/*')
+  app.route('/:url(api|auth|components|app|bower_components|assets|uploads|static)/*')
    .get(errors[404]);
 
   // All other routes should redirect to the index.html
